Data Visualization for the Forbs data using Tableau
The Data Visualisation Project consists of the following steps:
a) Review the dataset
b) Identify key data elements in the selected dataset that could be used to further explore using data visualisation methods.
c) Define a business/user persona that will be the audience of your data visualisation, and develop a story to help focus and guide your data visualisation development process.
d) Present a professional report outlining the key aspects of your data visualization.
e) Develop a Data Visualisation dashboard (using the Tableau software), and present screenshots and descriptions of the dashboard in PowerPoint.